milestone_1/infra/requirements.txt
pyyaml==6.0.2
types-PyYAML==6.0.2
aws-cdk-lib==2.180.0
constructs>=10.0.0,<11.0.0
pystache==0.6.8
cdk-ecr-deployment==3.0.23
aws-cdk-aws-lambda-python-alpha